requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML 라이브러리
torch==2.1.0
transformers==4.35.0
peft==0.6.0
accelerate==0.24.0

# 벡터 데이터베이스
chromadb==0.4.15

# OpenAI API
openai==1.3.0

# 환경 설정 및 유틸리티
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3

# 로깅 및 모니터링
loguru==0.7.2

# 개발 도구 (선택사항)
black==23.9.1
pytest==7.4.3
pytest-asyncio==0.21.1